Changed folder structure so that it can be used by another project by only copying the include directory and the shared library file -.so- that is built. Renamed Lanes.h and Lanes.cpp because the class declared was named Lane not Lanes. Updated CMakeLists.txt accordingly, built and tested succesfully.
