+++ CMakeLists.txt
@@ -12,34 +12,34 @@ set(CMAKE_CXX_FLAGS "-Wall")
 set(CMAKE_CXX_FLAGS_DEBUG "-g")
 set(CMAKE_CXX_FLAGS_RELEASE "-O3")
 
-include_directories(./)
+include_directories(include)
 include_directories(Thirdparty)
 
 add_library(OpenDrive SHARED
-    Geometries/Arc.cpp
-    Geometries/CubicSpline.cpp
-    Geometries/Line.cpp
-    Geometries/ParamPoly3.cpp
-    Geometries/RoadGeometry.cpp
-    Geometries/Spiral.cpp
-    Geometries/Spiral/odrSpiral.cpp
-    Lanes.cpp
-    LaneSection.cpp
-    Mesh.cpp
-    OpenDriveMap.cpp
-    RefLine.cpp
-    Road.cpp
-    RoadMark.cpp
-    RoadObject.cpp
-    RoutingGraph.cpp
+    src/Geometries/Arc.cpp
+    src/Geometries/CubicSpline.cpp
+    src/Geometries/Line.cpp
+    src/Geometries/ParamPoly3.cpp
+    src/Geometries/RoadGeometry.cpp
+    src/Geometries/Spiral.cpp
+    src/Geometries/Spiral/odrSpiral.cpp
+    src/Lane.cpp
+    src/LaneSection.cpp
+    src/Mesh.cpp
+    src/OpenDriveMap.cpp
+    src/RefLine.cpp
+    src/Road.cpp
+    src/RoadMark.cpp
+    src/RoadObject.cpp
+    src/RoutingGraph.cpp
     Thirdparty/pugixml/pugixml.cpp
 )
 
 if(EMSCRIPTEN)
     set(CMAKE_EXECUTABLE_SUFFIX ".js")
     add_executable(ModuleOpenDrive
-        Embind.cpp
-        Viewer/RoadNetworkMesh.cpp 
+        src/Embind.cpp
+        Viewer/RoadNetworkMesh.cpp
         Viewer/ViewerUtils.cpp
     )
     target_include_directories(ModuleOpenDrive PUBLIC Viewer)
@@ -56,6 +56,6 @@ if(EMSCRIPTEN)
 endif()
 
 if(NOT EMSCRIPTEN)
-    add_executable(test-xodr test.cpp)
+    add_executable(test-xodr src/test.cpp)
     target_link_libraries(test-xodr OpenDrive)
 endif()
