+++ include/Geometries/Spiral.h
