+++ include/XmlNode.h
