+++ src/OpenDriveMap.cpp
@@ -7,7 +7,7 @@
 #include "Geometries/Spiral.h"
 #include "Junction.h"
 #include "LaneSection.h"
-#include "Lanes.h"
+#include "Lane.h"
 #include "Math.hpp"
 #include "RefLine.h"
 #include "Road.h"
